Remove row-state backfill and diagnostic tooling
