+++ src/feature-flags/feature-flags.config.js
@@ -11,12 +11,6 @@ export const createConfigFeatureFlags = (config) => ({
   isSummaryLogRowStatesEnabled() {
     return config.get('featureFlags.summaryLogRowStates')
   },
-  isSummaryLogRowStatesBackfillEnabled() {
-    return config.get('featureFlags.summaryLogRowStatesBackfill')
-  },
-  isSummaryLogRowStatesDiscrepancyReportEnabled() {
-    return config.get('featureFlags.summaryLogRowStatesDiscrepancyReport')
-  },
   isRegisteredOnlySubmittedEventsEnabled() {
     return config.get('featureFlags.registeredOnlySubmittedEvents')
   },
