+++ src/feature-flags/feature-flags.inmemory.js
@@ -12,12 +12,6 @@ export const createInMemoryFeatureFlags = (flags = {}) => ({
   isSummaryLogRowStatesEnabled() {
     return flags.summaryLogRowStates ?? false
   },
-  isSummaryLogRowStatesBackfillEnabled() {
-    return flags.summaryLogRowStatesBackfill ?? false
-  },
-  isSummaryLogRowStatesDiscrepancyReportEnabled() {
-    return flags.summaryLogRowStatesDiscrepancyReport ?? false
-  },
   isRegisteredOnlySubmittedEventsEnabled() {
     return flags.registeredOnlySubmittedEvents ?? false
   },
