+++ src/feature-flags/feature-flags.port.js
@@ -3,8 +3,6 @@
  * @property {() => boolean} isDevEndpointsEnabled
  * @property {() => boolean} isFixDuplicateAccreditationLinksEnabled
  * @property {() => boolean} isSummaryLogRowStatesEnabled
- * @property {() => boolean} isSummaryLogRowStatesBackfillEnabled
- * @property {() => boolean} isSummaryLogRowStatesDiscrepancyReportEnabled
  * @property {() => boolean} isRegisteredOnlySubmittedEventsEnabled
  * @property {() => boolean} isStaleIssuedTonnageReportEnabled
  */
@@ -14,8 +12,6 @@
  * @property {boolean} [devEndpoints]
  * @property {boolean} [fixDuplicateAccreditationLinks]
  * @property {boolean} [summaryLogRowStates]
- * @property {boolean} [summaryLogRowStatesBackfill]
- * @property {boolean} [summaryLogRowStatesDiscrepancyReport]
  * @property {boolean} [registeredOnlySubmittedEvents]
  * @property {boolean} [staleIssuedTonnageReport]
  */
