+++ src/server/summary-log-row-states-registration.integration.test.js
@@ -4,7 +4,6 @@ import { randomUUID } from 'node:crypto'
 
 import { setupAuthContext } from '#vite/helpers/setup-auth-mocking.js'
 import { SUMMARY_LOG_ROW_STATES_COLLECTION_NAME } from '#waste-records/repository/mongodb.js'
-import { SUMMARY_LOG_ROW_STATES_BACKFILL_WATERMARKS_COLLECTION_NAME } from '#waste-records/backfill/watermark/mongodb.js'
 
 vi.mock(
   '#adapters/sqs-command-executor/sqs-command-executor.plugin.js',
@@ -52,7 +51,6 @@ describe('summary-log-row-states repository registration', () => {
 
     beforeAll(async () => {
       delete process.env.FEATURE_FLAG_SUMMARY_LOG_ROW_STATES
-      delete process.env.FEATURE_FLAG_SUMMARY_LOG_ROW_STATES_BACKFILL
       server = await bootServer()
     })
 
@@ -65,25 +63,9 @@ describe('summary-log-row-states repository registration', () => {
       expect(server.app.summaryLogRowStatesRepository).toBeUndefined()
     })
 
-    it('constructs no backfill watermark repository', () => {
-      expect(
-        server.app.summaryLogRowStatesBackfillWatermarkRepository
-      ).toBeUndefined()
-    })
-
     it('creates no row-state collection or indexes', async () => {
       const collections = await server.db
         .listCollections({ name: SUMMARY_LOG_ROW_STATES_COLLECTION_NAME })
-        .toArray()
-
-      expect(collections).toHaveLength(0)
-    })
-
-    it('creates no backfill watermark collection', async () => {
-      const collections = await server.db
-        .listCollections({
-          name: SUMMARY_LOG_ROW_STATES_BACKFILL_WATERMARKS_COLLECTION_NAME
-        })
         .toArray()
 
       expect(collections).toHaveLength(0)
@@ -95,7 +77,6 @@ describe('summary-log-row-states repository registration', () => {
 
     beforeAll(async () => {
       process.env.FEATURE_FLAG_SUMMARY_LOG_ROW_STATES = 'true'
-      delete process.env.FEATURE_FLAG_SUMMARY_LOG_ROW_STATES_BACKFILL
       server = await bootServer()
     })
 
@@ -107,24 +88,6 @@ describe('summary-log-row-states repository registration', () => {
 
     it('constructs the row-state repository', () => {
       expect(server.app.summaryLogRowStatesRepository).toBeDefined()
-    })
-
-    it('constructs the backfill watermark repository', () => {
-      expect(
-        server.app.summaryLogRowStatesBackfillWatermarkRepository
-      ).toBeDefined()
-    })
-
-    it('creates the backfill watermark collection with its unique registration index', async () => {
-      const indexes = await server.db
-        .collection(SUMMARY_LOG_ROW_STATES_BACKFILL_WATERMARKS_COLLECTION_NAME)
-        .listIndexes()
-        .toArray()
-
-      const registrationIdentity = indexes.find(
-        (index) => index.name === 'registration_identity'
-      )
-      expect(registrationIdentity.unique).toBe(true)
     })
 
     it('creates the empty collection with its three indexes', async () => {
